Modularise AI Lab models
