+++ app/modules/ai_lab/models/resource_listings.py
@@ -1,49 +1,12 @@
-from django import forms
 from django.db import models
 
 from wagtail.admin.edit_handlers import FieldPanel, StreamFieldPanel
-from wagtail.core.models import Page
 
 from modules.core.models.abstract import BasePage
 from modules.core.models.pages import SectionPage, ArticlePage
 from modules.blog_posts.models import BlogPost
 
-
-class AiLabHomePage(SectionPage):
-    subpage_types = ["AiLabResourceIndexPage", "core.ArticlePage"]
-    max_count = 1
-
-    def get_context(self, request):
-        context = super().get_context(request)
-        blog_posts = (
-            BlogPost.objects.filter(tags__slug="ai-lab")
-            .distinct()
-            .order_by("-first_published_at")[:3]
-        )
-        context.update({"blog_posts": blog_posts})
-        return context
-
-
-class AiLabResourceMixin(models.Model):
-    parent_page_types = [
-        "AiLabUnderstandIndexPage",
-        "AiLabDevelopIndexPage",
-        "AiLabAdoptIndexPage",
-    ]
-
-    class Meta:
-        abstract = True
-
-
-class AiLabCaseStudy(AiLabResourceMixin, ArticlePage):
-    pass
-
-
-class AiLabExternalResource(AiLabResourceMixin, Page):
-    external_url = models.URLField()
-    content_panels = Page.content_panels + [
-        FieldPanel("external_url", widget=forms.URLInput()),
-    ]
+from modules.ai_lab.models.resources import AiLabCaseStudy, AiLabExternalResource
 
 
 class AiLabResourceIndexPage(BasePage):
